+++ src/test/php/net/xp_framework/unittest/logging/LogCategoryTest.class.php
@@ -1,25 +1,18 @@
 <?php namespace net\xp_framework\unittest\logging;
  
-use unittest\TestCase;
+use util\log\LogCategory;
 use util\log\Logger;
 use util\log\Appender;
 use util\log\LogLevel;
+use util\log\LoggingEvent;
+use util\log\Context;
 use util\log\layout\PatternLayout;
 use util\log\context\NestedLogContext;
 
 /**
  * Tests LogCategory class
  */
-class LogCategoryTest extends TestCase {
-  public $cat= null;
-  
-  /**
-   * Setup method. Creates logger and cat member for easier access to
-   * the Logger instance
-   */
-  public function setUp() {
-    $this->cat= new \util\log\LogCategory('test');
-  }
+class LogCategoryTest extends \unittest\TestCase {
   
   /**
    * Create a mock appender which simply stores all messages passed to 
@@ -27,132 +20,176 @@ class LogCategoryTest extends TestCase {
    *
    * @return  util.log.Appender
    */
-  protected function mockAppender() {
-    $appender= newinstance('util.log.Appender', [], '{
-      public $messages= [];
-      
-      public function append(LoggingEvent $event) {
-        $this->messages[]= array(
+  private function mockAppender() {
+    $appender= newinstance('util.log.Appender', [], [
+      'messages' => [],
+      'append' => function(LoggingEvent $event) {
+        $this->messages[]= [
           strtolower(LogLevel::nameOf($event->getLevel())), 
           $this->layout->format($event)
-        );
+        ];
       }
-    }');
+    ]);
     return $appender->withLayout(new PatternLayout('%m'));
+  }
+
+  /**
+   * Create an empty appender
+   *
+   * @return  util.log.Appender
+   */
+  private function emptyAppender() {
+    return newinstance('util.log.Appender', [], [
+      'append' => function(LoggingEvent $event) { }
+    ]);
   }
 
   /**
    * Helper method
    *
-   * @param   string $method
-   * @param   var $func A function to be called with `$this->cat` as argument
+   * @param   var $expected
+   * @param   util.log.LogCategory $cat
+   * @param   function(util.log.LogCategory): void $func A function to be called with `$cat` as argument
    * @throws  unittest.AssertionFailedError
    */
-  protected function assertLogged($result, $func) {
-    $app= $this->cat->addAppender($this->mockAppender());
-    $func($this->cat);
+  private function assertLogged($result, $cat, $func) {
+    $app= $cat->addAppender($this->mockAppender());
+    $func($cat);
     $this->assertEquals($result, $app->messages);
   }
-  
+
+  #[@test]
+  public function can_create_without_arguments() {
+    new LogCategory();
+  }
+
+  #[@test]
+  public function can_create_with_identifier() {
+    new LogCategory('identifier');
+  }
+
+  #[@test]
+  public function can_create_with_identifier_and_level() {
+    new LogCategory('identifier', LogLevel::ALL);
+  }
+
+  #[@test]
+  public function can_create_with_identifier_level_and_context() {
+    new LogCategory('identifier', LogLevel::ALL, newinstance(Context::class, [], [
+      'format' => function() { return ''; }
+    ]));
+  }
+
+  #[@test]
+  public function identifier() {
+    $this->assertEquals('identifier', (new LogCategory('identifier'))->identifier);
+  }
+
+  #[@test]
+  public function identifier_defaults_to_default() {
+    $this->assertEquals('default', (new LogCategory())->identifier);
+  }
+
   #[@test]
   public function logger_category_initially_has_no_appenders() {
-    $this->assertFalse($this->cat->hasAppenders());
+    $this->assertFalse((new LogCategory())->hasAppenders());
   }
 
   #[@test]
   public function addAappender_returns_added_appender() {
     $appender= $this->mockAppender();
-    $this->assertEquals($appender, $this->cat->addAppender($appender));
+    $this->assertEquals($appender, (new LogCategory())->addAppender($appender));
   }
 
   #[@test]
   public function withAppender_returns_category() {
-    $this->assertEquals($this->cat, $this->cat->withAppender($this->mockAppender()));
-  }
-
-  #[@test]
-  public function hasAppenders_initially_returns_false() {
-    $this->assertFalse($this->cat->hasAppenders());
+    $cat= new LogCategory();
+    $this->assertEquals($cat, $cat->withAppender($this->mockAppender()));
   }
 
   #[@test]
   public function hasAppenders_returns_true_after_adding_an_appender() {
-    $this->cat->addAppender($this->mockAppender());
-    $this->assertTrue($this->cat->hasAppenders());
+    $cat= new LogCategory();
+    $cat->addAppender($this->mockAppender());
+    $this->assertTrue($cat->hasAppenders());
   }
 
   #[@test]
   public function hasAppenders_returns_false_after_removing_added_appender() {
-    $a= $this->cat->addAppender($this->mockAppender());
-    $this->cat->removeAppender($a);
-    $this->assertFalse($this->cat->hasAppenders());
+    $cat= new LogCategory();
+    $a= $cat->addAppender($this->mockAppender());
+    $cat->removeAppender($a);
+    $this->assertFalse($cat->hasAppenders());
   }
 
   #[@test]
   public function adding_appender_twice_with_same_flags_has_no_effect() {
+    $cat= new LogCategory();
     $a= $this->mockAppender();
-    $this->cat->addAppender($a);
-    $this->cat->addAppender($a);
-    $this->assertEquals(array($a), $this->cat->getAppenders());
+    $cat->addAppender($a);
+    $cat->addAppender($a);
+    $this->assertEquals([$a], $cat->getAppenders());
   }
 
   #[@test]
   public function adding_appender_twice_with_differing_flags() {
+    $cat= new LogCategory();
     $a= $this->mockAppender();
-    $this->cat->addAppender($a, \util\log\LogLevel::INFO);
-    $this->cat->addAppender($a, \util\log\LogLevel::WARN);
-    $this->cat->removeAppender($a, \util\log\LogLevel::INFO);
-    $this->assertTrue($this->cat->hasAppenders());
-    $this->cat->removeAppender($a, \util\log\LogLevel::WARN);
-    $this->assertFalse($this->cat->hasAppenders());
+    $cat->addAppender($a, LogLevel::INFO);
+    $cat->addAppender($a, LogLevel::WARN);
+    $cat->removeAppender($a, LogLevel::INFO);
+    $this->assertTrue($cat->hasAppenders());
+    $cat->removeAppender($a, LogLevel::WARN);
+    $this->assertFalse($cat->hasAppenders());
   }
 
   #[@test]
   public function addAppender_sets_layout_if_appender_does_not_have_layout() {
-    $appender= newinstance('util.log.Appender', [], '{
-      public function append(LoggingEvent $event) { }
-    }');
-    $this->cat->addAppender($appender);
+    $cat= new LogCategory();
+    $appender= $this->emptyAppender();
+    $cat->addAppender($appender);
     $this->assertInstanceOf('util.log.layout.DefaultLayout', $appender->getLayout());
   }
 
   #[@test]
   public function addAppender_does_not_overwrite_layout() {
-    $appender= newinstance('util.log.Appender', [], '{
-      public function append(LoggingEvent $event) { }
-    }');
-    $this->cat->addAppender($appender->withLayout(new PatternLayout('%m')));
+    $cat= new LogCategory();
+    $appender= $this->emptyAppender();
+    $cat->addAppender($appender->withLayout(new PatternLayout('%m')));
     $this->assertInstanceOf('util.log.layout.PatternLayout', $appender->getLayout());
   }
 
   #[@test]
   public function withAppender_sets_layout_if_appender_does_not_have_layout() {
-    $appender= newinstance('util.log.Appender', [], '{
-      public function append(LoggingEvent $event) { }
-    }');
-    $this->cat->withAppender($appender);
+    $cat= new LogCategory();
+    $appender= $this->emptyAppender();
+    $cat->withAppender($appender);
     $this->assertInstanceOf('util.log.layout.DefaultLayout', $appender->getLayout());
   }
 
   #[@test]
   public function withAppender_does_not_overwrite_layout() {
-    $appender= newinstance('util.log.Appender', [], '{
-      public function append(LoggingEvent $event) { }
-    }');
-    $this->cat->withAppender($appender->withLayout(new PatternLayout('%m')));
+    $cat= new LogCategory();
+    $appender= $this->emptyAppender();
+    $cat->withAppender($appender->withLayout(new PatternLayout('%m')));
     $this->assertInstanceOf('util.log.layout.PatternLayout', $appender->getLayout());
   }
 
   #[@test]
+  public function log_categories_with_empty_identifiers_are_equal() {
+    $this->assertEquals(new LogCategory(), new LogCategory());
+  }
+
+  #[@test]
   public function log_categories_with_same_identifiers_are_equal() {
-    $this->assertEquals(new \util\log\LogCategory('test'), $this->cat);
+    $this->assertEquals(new LogCategory('test'), new LogCategory('test'));
   }
 
   #[@test]
   public function log_categories_with_differing_appenders_are_not_equal() {
     $this->assertNotEquals(
-      new \util\log\LogCategory('test'), 
-      $this->cat->withAppender($this->mockAppender())
+      new LogCategory(),
+      (new LogCategory())->withAppender($this->mockAppender())
     );
   }
 
@@ -160,8 +197,8 @@ class LogCategoryTest extends TestCase {
   public function log_categories_with_appenders_differing_in_flags_are_not_equal() {
     $appender= $this->mockAppender();
     $this->assertNotEquals(
-      (new \util\log\LogCategory('test'))->withAppender($appender, \util\log\LogLevel::WARN), 
-      $this->cat->withAppender($appender)
+      (new LogCategory())->withAppender($appender, LogLevel::WARN), 
+      (new LogCategory())->withAppender($appender)
     );
   }
 
@@ -169,15 +206,16 @@ class LogCategoryTest extends TestCase {
   public function log_categories_with_same_appenders_are_equal() {
     $appender= $this->mockAppender();
     $this->assertEquals(
-      (new \util\log\LogCategory('test'))->withAppender($appender), 
-      $this->cat->withAppender($appender)
+      (new LogCategory())->withAppender($appender), 
+      (new LogCategory())->withAppender($appender)
     );
   }
 
   #[@test]
   public function debug() {
     $this->assertLogged(
-      array(array('debug', 'Test')),
+      [['debug', 'Test']],
+      new LogCategory(),
       function($cat) { $cat->debug('Test'); }
     );
   }
@@ -185,7 +223,8 @@ class LogCategoryTest extends TestCase {
   #[@test]
   public function debugf() {
     $this->assertLogged(
-      array(array('debug', 'Test 123')),
+      [['debug', 'Test 123']],
+      new LogCategory(),
       function($cat) { $cat->debugf('Test %d', '123'); }
     );
   }
@@ -193,7 +232,8 @@ class LogCategoryTest extends TestCase {
   #[@test]
   public function info() {
     $this->assertLogged(
-      array(array('info', 'Test')),
+      [['info', 'Test']],
+      new LogCategory(),
       function($cat) { $cat->info('Test'); }
     );
   }
@@ -201,7 +241,8 @@ class LogCategoryTest extends TestCase {
   #[@test]
   public function infof() {
     $this->assertLogged(
-      array(array('info', 'Test 123')),
+      [['info', 'Test 123']],
+      new LogCategory(),
       function($cat) { $cat->infof('Test %d', '123'); }
     );
   }
@@ -209,7 +250,8 @@ class LogCategoryTest extends TestCase {
   #[@test]
   public function warn() {
     $this->assertLogged(
-      array(array('warn', 'Test')),
+      [['warn', 'Test']],
+      new LogCategory(),
       function($cat) { $cat->warn('Test'); }
     );
   }
@@ -217,7 +259,8 @@ class LogCategoryTest extends TestCase {
   #[@test]
   public function warnf() {
     $this->assertLogged(
-      array(array('warn', 'Test 123')),
+      [['warn', 'Test 123']],
+      new LogCategory(),
       function($cat) { $cat->warnf('Test %d', '123'); }
     );
   }
@@ -225,7 +268,8 @@ class LogCategoryTest extends TestCase {
   #[@test]
   public function error() {
     $this->assertLogged(
-      array(array('error', 'Test')),
+      [['error', 'Test']],
+      new LogCategory(),
       function($cat) { $cat->error('Test'); }
     );
   }
@@ -233,7 +277,8 @@ class LogCategoryTest extends TestCase {
   #[@test]
   public function errorf() {
     $this->assertLogged(
-      array(array('error', 'Test 123')),
+      [['error', 'Test 123']],
+      new LogCategory(),
       function($cat) { $cat->errorf('Test %d', '123'); }
     );
   }
@@ -241,7 +286,8 @@ class LogCategoryTest extends TestCase {
   #[@test]
   public function mark() {
     $this->assertLogged(
-      array(array('info', str_repeat('-', 72))),
+      [['info', str_repeat('-', 72)]],
+      new LogCategory(),
       function($cat) { $cat->mark(); }
     );
   }
@@ -249,80 +295,90 @@ class LogCategoryTest extends TestCase {
   #[@test]
   public function log() {
     $this->assertLogged(
-      array(array('info', 'Test 123')),
-      function($cat) { $cat->log(LogLevel::INFO, array('Test', '123')); }
+      [['info', 'Test 123']],
+      new LogCategory(),
+      function($cat) { $cat->log(LogLevel::INFO, ['Test', '123']); }
     );
   }
 
   #[@test]
   public function warning_message_only_gets_appended_to_warn_appender() {
-    $app1= $this->cat->addAppender($this->mockAppender(), \util\log\LogLevel::INFO);
-    $app2= $this->cat->addAppender($this->mockAppender(), \util\log\LogLevel::WARN);
-    $this->cat->warn('Test');
+    $cat= new LogCategory();
+    $app1= $cat->addAppender($this->mockAppender(), LogLevel::INFO);
+    $app2= $cat->addAppender($this->mockAppender(), LogLevel::WARN);
+    $cat->warn('Test');
     $this->assertEquals([], $app1->messages);
-    $this->assertEquals(array(array('warn', 'Test')), $app2->messages); 
+    $this->assertEquals([['warn', 'Test']], $app2->messages);
   }
 
   #[@test]
   public function getAppenders_initially_returns_empty_array() {
-    $this->assertEquals([], $this->cat->getAppenders());
+    $this->assertEquals([], (new LogCategory())->getAppenders());
   }
 
   #[@test]
   public function getAppenders_returns_added_appender() {
+    $cat= new LogCategory();
     $appender= $this->mockAppender();
-    $this->cat->addAppender($appender);
-    $this->assertEquals(array($appender), $this->cat->getAppenders());
+    $cat->addAppender($appender);
+    $this->assertEquals([$appender], $cat->getAppenders());
   }
 
   #[@test]
   public function getAppenders_returns_added_appender_with_error_flags() {
-    $appender= $this->cat->addAppender($this->mockAppender(), \util\log\LogLevel::ERROR);
-    $this->assertEquals(array($appender), $this->cat->getAppenders());
+    $cat= new LogCategory();
+    $appender= $cat->addAppender($this->mockAppender(), LogLevel::ERROR);
+    $this->assertEquals([$appender], $cat->getAppenders());
   }
 
   #[@test]
   public function getAppenders_with_error_flags_returns_added_appender_with_error_flags() {
-    $appender= $this->cat->addAppender($this->mockAppender(), \util\log\LogLevel::ERROR);
-    $this->assertEquals(array($appender), $this->cat->getAppenders(\util\log\LogLevel::ERROR));
+    $cat= new LogCategory();
+    $appender= $cat->addAppender($this->mockAppender(), LogLevel::ERROR);
+    $this->assertEquals([$appender], $cat->getAppenders(LogLevel::ERROR));
   }
 
   #[@test]
   public function getAppenders_with_info_flags_does_not_return_added_appender_with_error_flags() {
-    $appender= $this->cat->addAppender($this->mockAppender(), \util\log\LogLevel::ERROR);
-    $this->assertEquals([], $this->cat->getAppenders(\util\log\LogLevel::INFO));
+    $cat= new LogCategory();
+    $appender= $cat->addAppender($this->mockAppender(), LogLevel::ERROR);
+    $this->assertEquals([], $cat->getAppenders(LogLevel::INFO));
   }
 
   #[@test]
   public function getAppenders_with_info_flags_returns_added_appender_with_info_and_error_flags() {
-    $appender= $this->cat->addAppender($this->mockAppender(), \util\log\LogLevel::ERROR | \util\log\LogLevel::INFO);
-    $this->assertEquals(array($appender), $this->cat->getAppenders(\util\log\LogLevel::INFO));
+    $cat= new LogCategory();
+    $appender= $cat->addAppender($this->mockAppender(), LogLevel::ERROR | LogLevel::INFO);
+    $this->assertEquals([$appender], $cat->getAppenders(LogLevel::INFO));
   }
 
   #[@test]
   public function getAppenders_returns_appenders_with_flags() {
-    $app1= $this->cat->addAppender($this->mockAppender(), \util\log\LogLevel::ERROR);
-    $app2= $this->cat->addAppender($this->mockAppender(), \util\log\LogLevel::WARN);
-    $app3= $this->cat->addAppender($this->mockAppender(), \util\log\LogLevel::INFO);
-    $app4= $this->cat->addAppender($this->mockAppender(), \util\log\LogLevel::DEBUG);
-    $this->assertEquals(array($app1, $app2, $app3, $app4), $this->cat->getAppenders());
+    $cat= new LogCategory();
+    $app1= $cat->addAppender($this->mockAppender(), LogLevel::ERROR);
+    $app2= $cat->addAppender($this->mockAppender(), LogLevel::WARN);
+    $app3= $cat->addAppender($this->mockAppender(), LogLevel::INFO);
+    $app4= $cat->addAppender($this->mockAppender(), LogLevel::DEBUG);
+    $this->assertEquals([$app1, $app2, $app3, $app4], $cat->getAppenders());
   }
 
   #[@test]
   public function hasContext_initially_returns_false() {
-    $this->assertFalse($this->cat->hasContext());
+    $this->assertFalse((new LogCategory())->hasContext());
   }
 
   #[@test]
   public function hasContext_returns_true_after_setting_context() {
-    $this->cat->setContext(new NestedLogContext());
-    $this->assertTrue($this->cat->hasContext());
+    $cat= new LogCategory();
+    $cat->setContext(new NestedLogContext());
+    $this->assertTrue($cat->hasContext());
   }
 
   #[@test]
   public function getContext_returns_context_previously_set_with_setContext() {
+    $cat= new LogCategory();
     $context= new NestedLogContext();
-    $this->cat->setContext($context);
-    $this->assertEquals($context, $this->cat->getContext());
+    $cat->setContext($context);
+    $this->assertEquals($context, $cat->getContext());
   }
 }